fixed clang-tidy data type conflicts (switched int to size_t)

+++ tasks/ovchinnikov_m_max_values_in_matrix_rows/common/include/common.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <string>
 #include <tuple>
 #include <vector>
@@ -8,9 +9,9 @@
 
 namespace ovchinnikov_m_max_values_in_matrix_rows {
 
-using InType = std::tuple<int, int, std::vector<int>>;
+using InType = std::tuple<size_t, size_t, std::vector<int>>;
 using OutType = std::vector<int>;
-using TestType = std::tuple<int, int, std::vector<int>, std::string>;
+using TestType = std::tuple<size_t, size_t, std::vector<int>, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace ovchinnikov_m_max_values_in_matrix_rows
